Update VisualizeDataCube
Now get column using click and save it into a file named currColumn to be fed to Hough transform. It also saves x coordinates of the center of the column for back transform to the original image. The y coordinate can be ignored as it is only the y-coordinate where the user clicked